refactor: rename Iceworks Material Helper to React Component Helper
