+++ extensions/iceworks-app/src/quickPicks/showExtensionsQuickPick.ts
@@ -6,7 +6,7 @@ const extensionOptions = [
   { label: 'Iceworks 创建应用', detail: '快速创建多端应用（例如：React/Rax/Vue...）', command: 'iceworks-project-creator.start', },
   { label: 'Iceworks 生成页面', detail: '使用低代码的方式生成网页视图', command: 'iceworks-page-builder.create', },
   { label: 'Iceworks 生成组件', detail: '使用低代码的方式生成前端组件', command: 'iceworks-component-builder.generate' },
-  { label: 'Iceworks 使用物料', detail: '使用可视化的方式添加物料到应用中', command: 'iceworks-material-import.start' },
+  { label: 'Iceworks 使用物料', detail: '使用可视化的方式添加物料到应用中', command: 'iceworks-material-helper.start' },
 ]
 
 export default function showExtensionsQuickPick() {
