+++ extensions/iceworks-material-helper/README.en.md
@@ -3,18 +3,47 @@ English | [简体中文](./README.md)
 [![Version](https://vsmarketplacebadge.apphb.com/version/iceworks-team.iceworks-material-helper.svg)](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-material-helper)
 [![Installs](https://vsmarketplacebadge.apphb.com/installs-short/iceworks-team.iceworks-material-helper.svg)](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-material-helper)
 
-# Iceworks Material Helper
+# React Component Helper
 
-Write props easier in [JSX](https://reactjs.org/docs/introducing-jsx.html), friendly for React and [Rax](https://rax.js.org/).
+Use Component and Write Props easier in [JSX](https://reactjs.org/docs/introducing-jsx.html), friendly for React and [Rax](https://rax.js.org/).
 
 ## Usage
+
+### Using Material
+
+#### Activate Material Panel
+
+![demo](https://user-images.githubusercontent.com/56879942/87401390-a63a1700-c5ec-11ea-91a6-f7d81ac5fd77.gif)
+
+1. Open vscode command palette  through `Ctrl+Shift+P` or `⇧⌘P`.
+2. Enter `Iceworks: Import Material` to activate Material Panel.
+
+#### Use in Pages
+
+![demo](https://user-images.githubusercontent.com/56879942/87619860-ba4a5980-c74f-11ea-84c1-9ef69ef17b18.gif)
+
+1. Click src/pages/*/index.tsx in the resource panel to open a page.
+2. Activate the Material Panel.
+3. Move cursor to a position for materials.
+4. Search the block materials, click the material to be used, and the material code will be automatically added to the corresponding position.
+5. Search for component materials. Click the material to be used, and the material code will be automatically added to the corresponding position.
+
+#### Use in Components
+
+![demo](https://user-images.githubusercontent.com/56879942/87619875-c2a29480-c74f-11ea-945e-788a32e65881.gif)
+
+1. Click src/components/*/index.tsx in the resource panel to open a page.
+2. Activate the material-import extension.
+3. Move cursor to a position for materials.
+4. Search for component materials. Click the material to be used, and the material code will be automatically added to the corresponding position.
+
+### Props Autocomplete
 
 When editing the props of a component in a JSX file, an automatic completion reminder will be given:
 
 ![demo](https://user-images.githubusercontent.com/56879942/87399599-2dd25680-c5ea-11ea-9402-5e36ba7b8f98.gif)
 
 1. Use materials in JSX files (such as View). 
-    > What is [material](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-component-builder)?
 2. Enter part of props in the material tag, it will arouse automatic completion.
 
 ## More
