+++ extensions/iceworks-material-helper/README.md
@@ -3,21 +3,50 @@
 [![Version](https://vsmarketplacebadge.apphb.com/version/iceworks-team.iceworks-material-helper.svg)](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-material-helper)
 [![Installs](https://vsmarketplacebadge.apphb.com/installs-short/iceworks-team.iceworks-material-helper.svg)](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-material-helper)
 
-# Iceworks 物料使用辅助插件
+# React 组件开发辅助插件
 
-在 [JSX](https://zh-hans.reactjs.org/docs/introducing-jsx.html) 中更快速地编写组件属性，对 React 和 [Rax](https://rax.js.org/) 应用友好。
+在 [JSX](https://zh-hans.reactjs.org/docs/introducing-jsx.html) 中更快更好地添加组件、编写组件属性，对 React 和 [Rax](https://rax.js.org/) 应用友好。
 
 ## 功能
 
-### 属性自动补全
+### 使用物料
+
+> [什么是物料？](https://ice.work/docs/materials/about)
+
+#### 激活物料面板
+
+![使用示例](https://user-images.githubusercontent.com/56879942/87538941-a19a5f00-c6cf-11ea-92f2-b8ed100792fc.gif)
+
+1. 通过 `⇧⌘P` 或 `Ctrl+Shift+P` 快捷键唤醒命令面板
+2. 输入 `Iceworks: 使用物料` 激活物料添加面板
+
+#### 在页面中使用物料
+
+![使用示例](https://user-images.githubusercontent.com/56879942/87619860-ba4a5980-c74f-11ea-84c1-9ef69ef17b18.gif)
+
+1. 在资源面板中点击 src/pages/*/index.tsx 打开页面源码
+2. 激活物料添加面板
+3. 选择需要插入物料的代码位置
+4. 搜索区块，点击需要使用的区块，物料代码将自动添加到相应的位置中
+5. 搜索组件，点击需要使用的组件，物料代码将自动添加到相应的位置中
+
+#### 在组件中使用物料
+
+![使用示例](https://user-images.githubusercontent.com/56879942/87619875-c2a29480-c74f-11ea-945e-788a32e65881.gif)
+
+1. 在资源面板中点击 src/components/*/index.tsx 打开页面源码
+2. 激活物料添加面板
+3. 选择需要插入物料的代码位置
+4. 搜索组件，点击需要使用的组件，物料代码将自动添加到相应的位置中
+
+### 组件属性自动补全
 
 在 JSX 文件中编辑组件的属性时将给予自动补全提醒：
 
 ![使用说明](https://user-images.githubusercontent.com/56879942/87399599-2dd25680-c5ea-11ea-9402-5e36ba7b8f98.gif)
 
-1. 在 JSX 文件中使用物料（如 View ）
-    > 什么是[物料](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-component-builder)？
-2. 在物料标签内输入属性的部分内容，将会激活自动补全提醒。
+1. 在 JSX 文件中使用组件（如 View ）
+2. 在组件标签内输入属性的部分内容，将会激活自动补全提醒。
 
 ## 更多
