+++ extensions/iceworks-material-helper/package.nls.json
@@ -1,0 +1,3 @@
+{
+  "iceworksMaterialHelper.commands.start.title": "Iceworks: Import Material"
+}
