+++ extensions/iceworks-material-helper/package.nls.zh-cn.json
@@ -1,0 +1,3 @@
+{
+  "iceworksMaterialHelper.commands.start.title": "Iceworks: 使用物料"
+}
