--- extensions/iceworks-material-helper/src/index.ts
@@ -1,8 +1,0 @@
-import * as vscode from 'vscode';
-import propsAutoComplete from './propsAutoComplete';
-
-function activate() {
-  propsAutoComplete();
-};
-
-exports.activate = activate;
