+++ extensions/iceworks-material-helper/tsconfig.json
@@ -5,7 +5,11 @@
     "rootDir": "src",
     "noImplicitAny": false
   },
+	"include": [
+		"src"
+	],
   "exclude": [
+    ".vscode-test",
     "node_modules"
   ]
 }
