--- extensions/iceworks-material-import/.vscodeignore
@@ -1,8 +1,0 @@
-.vscode/**
-.vscode-test/**
-**/tsconfig.json
-**/*.map
-**/*.ts
-web/**
-node_modules
-!node_modules/prettier
