--- extensions/iceworks-material-import/CHANGELOG.md
@@ -1,17 +1,0 @@
-# Change Log
-
-## 0.1.3
-
-- docs: change logo
-
-## 0.1.2
-
-- feat: quick activate iceworks-material-import extensions by editor context
-
-## 0.1.1
-
-- style: better style of the webview
-
-## 0.1.0
-
-- Initial release
