--- extensions/iceworks-material-import/README.en.md
@@ -1,44 +1,0 @@
-English | [简体中文](./README.md)
-
-[![Version](https://vsmarketplacebadge.apphb.com/version/iceworks-team.iceworks-material-import.svg)](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-material-import)
-[![Installs](https://vsmarketplacebadge.apphb.com/installs-short/iceworks-team.iceworks-material-import.svg)](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-material-import)
-
-# Iceworks Material Import
-
-Add materials to the application in a visual way.
-
-## Usage
-
-### Activate Extension
-
-![demo](https://user-images.githubusercontent.com/56879942/87401390-a63a1700-c5ec-11ea-91a6-f7d81ac5fd77.gif)
-
-1. Open vscode command palette  through `Ctrl+Shift+P` or `⇧⌘P`.
-2. Enter `Iceworks:Import Material` to activate this extension.
-
-### Use in pages
-
-![demo](https://user-images.githubusercontent.com/56879942/87619860-ba4a5980-c74f-11ea-84c1-9ef69ef17b18.gif)
-
-1. Click src/pages/*/index.tsx in the resource panel to open a page.
-2. Activate the material-import extension.
-3. Move cursor to a position for materials
-4. Search the block materials, click the material to be used, and the material code will be automatically added to the corresponding position.
-5. Search for component materials. Click the material to be used, and the material code will be automatically added to the corresponding position.
-
-### Use in components
-
-![demo](https://user-images.githubusercontent.com/56879942/87619875-c2a29480-c74f-11ea-945e-788a32e65881.gif)
-
-1. Click src/components/*/index.tsx in the resource panel to open a page.
-2. Activate the material-import extension.
-3. Move cursor to a position for materials
-4. Search for component materials. Click the material to be used, and the material code will be automatically added to the corresponding position.
-
-## More
-
-See [Iceworks Pack](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks) to know more features.
-
-## License
-
-[MIT](https://github.com/ice-lab/iceworks/blob/master/LICENSE)
