--- extensions/iceworks-material-import/README.md
@@ -1,44 +1,0 @@
-简体中文 | [English](./README.en.md)
-
-[![Version](https://vsmarketplacebadge.apphb.com/version/iceworks-team.iceworks-material-import.svg)](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-material-import)
-[![Installs](https://vsmarketplacebadge.apphb.com/installs-short/iceworks-team.iceworks-material-import.svg)](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-material-import)
-
-# Iceworks 物料添加插件
-
-使用可视化的方式添加物料到应用中。
-
-## 功能
-
-### 激活
-
-![使用示例](https://user-images.githubusercontent.com/56879942/87538941-a19a5f00-c6cf-11ea-92f2-b8ed100792fc.gif)
-
-1. 通过 `⇧⌘P` 或 `Ctrl+Shift+P` 快捷键唤醒命令面板
-2. 输入 `Iceworks:Import Material` 激活物料添加插件
-
-### 在页面中使用
-
-![使用示例](https://user-images.githubusercontent.com/56879942/87619860-ba4a5980-c74f-11ea-84c1-9ef69ef17b18.gif)
-
-1. 在资源面板中点击 src/pages/*/index.tsx 打开页面源码
-2. 激活物料添加面板
-3. 选择需要插入物料的代码位置
-4. 搜索区块，点击需要使用的区块，物料代码将自动添加到相应的位置中
-5. 搜索组件，点击需要使用的组件，物料代码将自动添加到相应的位置中
-
-### 在组件中使用
-
-![使用示例](https://user-images.githubusercontent.com/56879942/87619875-c2a29480-c74f-11ea-945e-788a32e65881.gif)
-
-1. 在资源面板中点击 src/components/*/index.tsx 打开页面源码
-2. 激活物料添加面板
-3. 选择需要插入物料的代码位置
-4. 搜索组件，点击需要使用的组件，物料代码将自动添加到相应的位置中
-
-## 更多
-
-访问 [Iceworks 套件](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks)获取更多功能。
-
-## License
-
-[MIT](https://github.com/ice-lab/iceworks/blob/master/LICENSE)
