--- extensions/iceworks-material-import/package.json
@@ -1,75 +1,0 @@
-{
-  "name": "iceworks-material-import",
-  "displayName": "Iceworks Material Importer",
-  "description": "Quick import material in a visual way.",
-  "publisher": "iceworks-team",
-  "version": "0.1.3",
-  "engines": {
-    "vscode": "^1.41.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ice-lab/iceworks.git"
-  },
-  "categories": [
-    "Other"
-  ],
-  "keywords": [
-    "ICE",
-    "Rax",
-    "React",
-    "Low-code",
-    "lowcode",
-    "visual"
-  ],
-  "icon": "assets/logo.png",
-  "activationEvents": [
-    "onCommand:iceworks-material-import.start"
-  ],
-  "main": "./build/extension.js",
-  "contributes": {
-    "commands": [
-      {
-        "command": "iceworks-material-import.start",
-        "title": "%iceworksMaterialImport.commands.start.title%"
-      }
-    ],
-    "menus": {
-      "editor/context": [
-        {
-          "command": "iceworks-material-import.start",
-          "group": "iceworksMaterialImport",
-          "when": "editorTextFocus && iceworks:isJSXFile"
-        }
-      ]
-    }
-  },
-  "scripts": {
-    "vscode:prepublish": "rm -rf build && npm run build:web && webpack --mode production",
-    "build:web": "cd web && npm run build",
-    "watch:web": "cd web && npm run start",
-    "webpack": "webpack --mode development",
-    "webpack-dev": "webpack --mode development --watch",
-    "compile": "tsc -p ./tsconfig.json"
-  },
-  "devDependencies": {
-    "@types/node": "^13.11.0",
-    "@types/vscode": "^1.41.0",
-    "ts-loader": "^7.0.5",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11"
-  },
-  "dependencies": {
-    "@iceworks/block-service": "^0.1.7",
-    "@iceworks/common-service": "^0.1.3",
-    "@iceworks/component-service": "^0.1.0",
-    "@iceworks/material-service": "^0.1.3",
-    "@iceworks/page-service": "^0.1.5",
-    "@iceworks/vscode-webview": "^0.1.1"
-  },
-  "homepage": "https://github.com/ice-lab/iceworks/blob/master/extensions/iceworks-material-import/README.md",
-  "bugs": {
-    "url": "https://github.com/ice-lab/iceworks/issues",
-    "email": "iceworksteam@163.com"
-  }
-}
