--- extensions/iceworks-material-import/package.nls.json
@@ -1,3 +1,0 @@
-{
-  "iceworksMaterialImport.commands.start.title": "Iceworks: Import Material"
-}
