--- extensions/iceworks-material-import/package.nls.zh-cn.json
@@ -1,3 +1,0 @@
-{
-  "iceworksMaterialImport.commands.start.title": "Iceworks: 使用物料"
-}
