--- extensions/iceworks-material-import/src/types.d.ts
@@ -1,6 +1,0 @@
-// TODO hack for build
-declare module 'vfile-message' {
-  // eslint-disable-next-line
-  interface VFileMessage {
-  }
-}
