--- extensions/iceworks-material-import/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"extends": "../../tsconfig.extension.json",
-	"compilerOptions": {
-		"rootDir": "./",
-		"outDir": "build"
-	},
-	"include": [
-		"src"
-	],
-	"exclude": [
-		".vscode-test",
-		"web"
-	]
-}
