+++ extensions/iceworks/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.3.0
+
+- remove: add [iceworks-material-import](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-material-import)
+
 ## 0.2.0
 
 - docs: new logo and document
@@ -11,7 +15,7 @@
 ## 0.1.6
 
 - feat: add [iceworks-config-helper](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-config-helper)
-- feat: add [iceworks-material-import](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-config-helper)
+- feat: add [iceworks-material-import](https://marketplace.visualstudio.com/items?itemName=iceworks-team.iceworks-material-import)
 
 ## 0.1.5
